+++ CHANGELOG.md
@@ -23,6 +23,7 @@ FEATURES:
 * Add command to bootstrap single VMs
 * Add command to bootstrap every VM of a vApp
 * Add command to create/revert a vApp snapshot
+* Add command to upload OVF
 
 CHANGES:
 * Renamed & enhanced command _vapp config network_ to _vapp network external_